Migrate forge zarith to forging file, add unit tests
